Added Temperature Range Safe Guard
Added checks to ignore the commands from android and turn off the heater
and on the air conditioner at 90F.  It also turns on the heater at 45F.
This is to ensure the house maintains a safe temperature range even if
the Android application locks up.  It is currently only working if the
app starts and initializes the analog pin for the temperature sensor.