Fixed Bugs
Added new method to ADC to fetch the analog reading and am now calling
it from thermostat.setTemperature()

Added check for zero so the safety system doesn't immediately turn on
the heater on boot

Moved call for temperature in the main loop so it is still called when
Android is disconnected.

Bugs:
It is not yet ignoring android requests when in override mode
You must connect an android device initially to start analog poling.