Fixed remaining bugs so that safeguards are active
Added code to poll thermostat after phone disconnects
Added code to ignore request from Android when in override mode
Added sleep function after activating and deactivating the safety
override to prevent rapid toggling due to temperature fluxuations.
Added code to allow port 0 to still behave when in override mode for the
on board LED.
Added a check for 0 steps result in the temperature check so it doesn't
turn on the heater on bootup.