Added initialization delay
With the addition of the low-pass filter on the Rev-B hardware, it takes
a second after power up in order to read the correct temperature
voltage.  This is enough time for the safety backup system to read a low
temperature and kick in.  I added a new "armed" variable so the backup
system doesn't turn on until it reaches operating temperature for the
first time.