Added Heartbeat
Added a new failsafe to handle issues related to loosing communication
with Android while not registering as a disconnect.  The firmware now
expects a pin to be toggled once per second as a heartbeat signal.  If
this heart beat isn't received for approximately 10 seconds, the
firmware will issue a hardreset in an attempt to re-establish
connection.